Progress with pointers
Using two arrays to track and sort the frequency table while building a
tree with a third re-arranging the pointers. This got around several bad
exits prior (segfaults, bus errors and such). It compiles and runs to
completion but the binary trees are all zero. Was able to confirm that
the pointers in the lut map to the corresponding frequency table, but
the new->counts are all 0.